Address PR feedback: move test to unit tests and add news fragment
Move test_has_conflicting_changes_no_false_positive from component tests
to unit tests for faster execution and multi-Python-version coverage.
Save and restore registry._default_branch and repositories_directory
in a try/finally to prevent state leakage between tests.

Add changelog fragment for issue #8848.
