+++ changelog/8848.fixed.md
@@ -1,0 +1,1 @@
+Fix false-positive conflict detection during git sync caused by naive string matching of conflict markers in file content.
